+++ grakn-client/src/main/java/ai/grakn/remote/concept/RemoteAttribute.java
@@ -23,7 +23,7 @@ import ai.grakn.concept.AttributeType;
 import ai.grakn.concept.Concept;
 import ai.grakn.concept.ConceptId;
 import ai.grakn.concept.Thing;
-import ai.grakn.grpc.ConceptMethod;
+import ai.grakn.grpc.ConceptMethods;
 import ai.grakn.remote.RemoteGraknTx;
 import com.google.auto.value.AutoValue;
 
@@ -43,17 +43,17 @@ abstract class RemoteAttribute<D> extends RemoteThing<Attribute<D>, AttributeTyp
 
     @Override
     public final D getValue() {
-        return (D) runMethod(ConceptMethod.GET_VALUE);
+        return (D) runMethod(ConceptMethods.GET_VALUE);
     }
 
     @Override
     public final AttributeType.DataType<D> dataType() {
-        return (AttributeType.DataType<D>) runMethod(ConceptMethod.GET_DATA_TYPE_OF_ATTRIBUTE);
+        return (AttributeType.DataType<D>) runMethod(ConceptMethods.GET_DATA_TYPE_OF_ATTRIBUTE);
     }
 
     @Override
     public final Stream<Thing> ownerInstances() {
-        return runMethod(ConceptMethod.GET_OWNERS).map(Concept::asThing);
+        return runMethod(ConceptMethods.GET_OWNERS).map(Concept::asThing);
     }
 
     @Override
