+++ grakn-client/src/main/java/ai/grakn/remote/concept/RemoteRole.java
@@ -23,7 +23,7 @@ import ai.grakn.concept.ConceptId;
 import ai.grakn.concept.RelationshipType;
 import ai.grakn.concept.Role;
 import ai.grakn.concept.Type;
-import ai.grakn.grpc.ConceptMethod;
+import ai.grakn.grpc.ConceptMethods;
 import ai.grakn.remote.RemoteGraknTx;
 import com.google.auto.value.AutoValue;
 
@@ -41,12 +41,12 @@ abstract class RemoteRole extends RemoteSchemaConcept<Role> implements Role {
 
     @Override
     public final Stream<RelationshipType> relationshipTypes() {
-        return runMethod(ConceptMethod.GET_RELATIONSHIP_TYPES_THAT_RELATE_ROLE).map(Concept::asRelationshipType);
+        return runMethod(ConceptMethods.GET_RELATIONSHIP_TYPES_THAT_RELATE_ROLE).map(Concept::asRelationshipType);
     }
 
     @Override
     public final Stream<Type> playedByTypes() {
-        return runMethod(ConceptMethod.GET_TYPES_THAT_PLAY_ROLE).map(Concept::asType);
+        return runMethod(ConceptMethods.GET_TYPES_THAT_PLAY_ROLE).map(Concept::asType);
     }
 
     @Override
