+++ grakn-grpc/src/main/java/ai/grakn/grpc/ConceptResponseType.java
@@ -21,20 +21,21 @@ package ai.grakn.grpc;
 import ai.grakn.concept.AttributeType;
 import ai.grakn.concept.Concept;
 import ai.grakn.concept.Label;
-import ai.grakn.concept.Role;
-import ai.grakn.concept.Thing;
 import ai.grakn.graql.Pattern;
 import ai.grakn.rpc.generated.GrpcConcept;
 import ai.grakn.rpc.generated.GrpcConcept.ConceptResponse;
+import ai.grakn.rpc.generated.GrpcGrakn.TxResponse;
+import ai.grakn.rpc.generated.GrpcIterator.IteratorId;
+import org.apache.tinkerpop.gremlin.util.function.TriConsumer;
+import org.apache.tinkerpop.gremlin.util.function.TriFunction;
 
 import javax.annotation.Nullable;
-import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 import java.util.function.BiConsumer;
 import java.util.function.BiFunction;
 import java.util.function.Function;
 import java.util.stream.Stream;
+import java.util.stream.StreamSupport;
 
 /**
  * Wrapper around the different types of responses to {@link ConceptMethod}s applied on {@link Concept}s.
@@ -61,15 +62,17 @@ abstract class ConceptResponseType<T> {
             (builder, val) -> builder.setOptionalConcept(GrpcUtil.convertOptionalConcept(val))
     );
 
-    public static final ConceptResponseType<Stream<? extends Concept>> CONCEPTS = ConceptResponseType.create(
-            (converter, response) -> GrpcUtil.convert(converter, response.getConcepts()),
-            (builder, val) -> builder.setConcepts(GrpcUtil.convert(val))
-    );
+    public static final ConceptResponseType<Stream<? extends Concept>> CONCEPTS =
+            ConceptResponseType.createStreamable(
+                    (converter, response) -> converter.convert(response.getConcept()),
+                    GrpcUtil::conceptResponse
+            );
 
-    public static final ConceptResponseType<Map<Role, Set<Thing>>> ROLE_PLAYERS = ConceptResponseType.create(
-            (converter, response) -> GrpcUtil.convert(converter, response.getRolePlayers()),
-            (builder, val) -> builder.setRolePlayers(GrpcUtil.convert(val))
-    );
+    public static final ConceptResponseType<Stream<? extends RolePlayer>> ROLE_PLAYERS =
+            ConceptResponseType.createStreamable(
+                    (converter, response) -> converter.convert(response.getRolePlayer()),
+                    GrpcUtil::rolePlayerResponse
+            );
 
     public static final ConceptResponseType<AttributeType.DataType<?>> DATA_TYPE = ConceptResponseType.create(
             response -> GrpcUtil.convert(response.getDataType()),
@@ -106,9 +109,9 @@ abstract class ConceptResponseType<T> {
     );
 
     @Nullable
-    public abstract T get(GrpcConceptConverter converter, ConceptResponse conceptResponse);
+    public abstract T get(GrpcConceptConverter converter, GrpcClient client, ConceptResponse conceptResponse);
 
-    public abstract void set(ConceptResponse.Builder builder, @Nullable T value);
+    public abstract void set(ConceptResponse.Builder builder, GrpcIterators iterators, @Nullable T value);
 
     public static <T> ConceptResponseType<T> create(
             Function<ConceptResponse, T> getter,
@@ -121,15 +124,50 @@ abstract class ConceptResponseType<T> {
             BiFunction<GrpcConceptConverter, ConceptResponse, T> getter,
             BiConsumer<ConceptResponse.Builder, T> setter
     ) {
+        return create(
+                (converter, client, response) -> getter.apply(converter, response),
+                (builder, iterators, val) -> setter.accept(builder, val)
+        );
+    }
+
+    public static <T> ConceptResponseType<Stream<? extends T>> createStreamable(
+            BiFunction<GrpcConceptConverter, TxResponse, T> getter,
+            Function<T, TxResponse> setter
+    ) {
+        return create(
+                (converter, client, response) -> {
+                    IteratorId iteratorId = response.getIteratorId();
+
+                    Iterable<T> iterable = () -> new GraknGrpcIterator<T>(client, iteratorId) {
+                        @Override
+                        protected T getNextFromResponse(TxResponse response) {
+                            return getter.apply(converter, response);
+                        }
+                    };
+
+                    return StreamSupport.stream(iterable.spliterator(), false);
+                },
+                (builder, iterators, val) -> {
+                    Stream<TxResponse> responses = val.map(setter);
+                    IteratorId iteratorId = iterators.add(responses.iterator());
+                    builder.setIteratorId(iteratorId);
+                }
+        );
+    }
+
+    public static <T> ConceptResponseType<T> create(
+            TriFunction<GrpcConceptConverter, GrpcClient, ConceptResponse, T> getter,
+            TriConsumer<ConceptResponse.Builder, GrpcIterators, T> setter
+    ) {
         return new ConceptResponseType<T>() {
             @Override
-            public T get(GrpcConceptConverter converter, ConceptResponse conceptResponse) {
-                return getter.apply(converter, conceptResponse);
+            public T get(GrpcConceptConverter converter, GrpcClient client, ConceptResponse conceptResponse) {
+                return getter.apply(converter, client, conceptResponse);
             }
 
             @Override
-            public void set(ConceptResponse.Builder builder, @Nullable T value) {
-                setter.accept(builder, value);
+            public void set(ConceptResponse.Builder builder, GrpcIterators iterators, @Nullable T value) {
+                setter.accept(builder, iterators, value);
             }
         };
     }
