+++ grakn-grpc/src/main/java/ai/grakn/grpc/GrpcClient.java
@@ -16,7 +16,7 @@
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
  */
 
-package ai.grakn.remote;
+package ai.grakn.grpc;
 
 import ai.grakn.concept.AttributeType;
 import ai.grakn.concept.Concept;
@@ -28,19 +28,14 @@ import ai.grakn.graql.Query;
 import ai.grakn.graql.Var;
 import ai.grakn.graql.admin.Answer;
 import ai.grakn.graql.internal.query.QueryAnswer;
-import ai.grakn.grpc.ConceptMethod;
-import ai.grakn.grpc.GrpcConceptConverter;
-import ai.grakn.grpc.GrpcUtil;
 import ai.grakn.grpc.GrpcUtil.ErrorType;
-import ai.grakn.grpc.TxGrpcCommunicator;
 import ai.grakn.grpc.TxGrpcCommunicator.Response;
 import ai.grakn.rpc.generated.GraknGrpc;
 import ai.grakn.rpc.generated.GrpcGrakn;
-import ai.grakn.rpc.generated.GrpcGrakn.IteratorId;
 import ai.grakn.rpc.generated.GrpcGrakn.TxRequest;
 import ai.grakn.rpc.generated.GrpcGrakn.TxResponse;
+import ai.grakn.rpc.generated.GrpcIterator.IteratorId;
 import ai.grakn.util.CommonUtil;
-import com.google.common.collect.AbstractIterator;
 import com.google.common.collect.ImmutableMap;
 import io.grpc.Metadata;
 import io.grpc.Status;
@@ -51,18 +46,19 @@ import javax.annotation.Nullable;
 import java.util.Iterator;
 import java.util.Optional;
 import java.util.stream.Stream;
+import java.util.stream.StreamSupport;
 
 /**
  * Communicates with a Grakn gRPC server, translating requests and responses to and from their gRPC representations.
  *
  * <p>
  *     This class is a light abstraction layer over gRPC - it understands how the sequence of calls should execute and
- *     how to translate gRPC objects into Java objects and back. However, any logic is kept in {@link RemoteGraknTx}.
+ *     how to translate gRPC objects into Java objects and back.
  * </p>
  *
  * @author Felix Chapman
  */
-public final class GrpcClient implements AutoCloseable {
+public class GrpcClient implements AutoCloseable {
 
     private final GrpcConceptConverter conceptConverter;
     private final TxGrpcCommunicator communicator;
@@ -82,27 +78,15 @@ public final class GrpcClient implements AutoCloseable {
         responseOrThrow();
     }
 
-    public Iterator<Object> execQuery(RemoteGraknTx tx, Query<?> query) {
+    public Iterator<Object> execQuery(Query<?> query) {
         communicator.send(GrpcUtil.execQueryRequest(query.toString(), query.inferring()));
 
         IteratorId iteratorId = responseOrThrow().getIteratorId();
 
-        return new AbstractIterator<Object>() {
+        return new GraknGrpcIterator<Object>(this, iteratorId) {
             @Override
-            protected Object computeNext() {
-                communicator.send(GrpcUtil.nextRequest(iteratorId));
-
-                TxResponse response = responseOrThrow();
-
-                switch (response.getResponseCase()) {
-                    case QUERYRESULT:
-                        return convert(response.getQueryResult());
-                    case DONE:
-                        return endOfData();
-                    default:
-                    case RESPONSE_NOT_SET:
-                        throw CommonUtil.unreachableStatement("Unexpected " + response);
-                }
+            protected Object getNextFromResponse(TxResponse response) {
+                return convert(response.getQueryResult());
             }
         };
     }
@@ -112,10 +96,15 @@ public final class GrpcClient implements AutoCloseable {
         responseOrThrow();
     }
 
+    public TxResponse next(IteratorId iteratorId) {
+        communicator.send(GrpcUtil.nextRequest(iteratorId));
+        return responseOrThrow();
+    }
+
     @Nullable
     public <T> T runConceptMethod(ConceptId id, ConceptMethod<T> conceptMethod) {
         communicator.send(GrpcUtil.runConceptMethodRequest(id, conceptMethod));
-        return conceptMethod.get(conceptConverter, responseOrThrow());
+        return conceptMethod.get(conceptConverter, this, responseOrThrow());
     }
 
     public Optional<Concept> getConcept(ConceptId id) {
@@ -130,7 +119,17 @@ public final class GrpcClient implements AutoCloseable {
 
     public Stream<? extends Concept> getAttributesByValue(Object value) {
         communicator.send(GrpcUtil.getAttributesByValueRequest(value));
-        return GrpcUtil.convert(conceptConverter, responseOrThrow().getConcepts());
+
+        IteratorId iteratorId = responseOrThrow().getIteratorId();
+
+        Iterable<Concept> iterable = () -> new GraknGrpcIterator<Concept>(this, iteratorId) {
+            @Override
+            protected Concept getNextFromResponse(TxResponse response) {
+                return conceptConverter.convert(response.getConcept());
+            }
+        };
+
+        return StreamSupport.stream(iterable.spliterator(), false);
     }
 
     public Concept putEntityType(Label label) {
@@ -225,4 +224,5 @@ public final class GrpcClient implements AutoCloseable {
 
         return new QueryAnswer(map.build());
     }
+
 }
